Extract Parts in its own file
