+++ src/refactorings/extract/extract-variable/occurrence.ts
@@ -9,6 +9,7 @@ import {
   MemberExpressionVariable,
   ShorthandVariable
 } from "./variable";
+import { Parts } from "./parts";
 
 export { createOccurrence, Occurrence };
 
@@ -257,31 +258,3 @@ class PartialTemplateLiteralOccurrence extends Occurrence<t.TemplateLiteral> {
 }
 
 type IndentationLevel = number;
-
-class Parts {
-  constructor(
-    private readonly code: Code,
-    private readonly selection: Selection,
-    private readonly offset: Position
-  ) {}
-
-  get left(): Code {
-    return this.code.slice(0, this.start);
-  }
-
-  get value(): Code {
-    return this.code.slice(this.start, this.end);
-  }
-
-  get right(): Code {
-    return this.code.slice(this.end);
-  }
-
-  private get start(): number {
-    return this.selection.start.character - this.offset.character;
-  }
-
-  private get end(): number {
-    return this.selection.end.character - this.offset.character;
-  }
-}
